Fixes and libftgod

+++ Libft/ft_memchr.c
@@ -6,16 +6,16 @@
 /*   By: audgiova <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/08 16:19:44 by audgiova          #+#    #+#             */
-/*   Updated: 2026/06/11 17:27:17 by audgiova         ###   ########.fr       */
+/*   Updated: 2026/06/14 20:52:07 by audgiova         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char	*ptr;
-	int				i;
+	size_t			i;
 
 	i = 0;
 	ptr = (unsigned char *)s;
